requirements.txt
RNS==1.0.2
LXMF==0.9.2
SQLAlchemy
jsonschema
fastapi
httpx
msgpack
python-dotenv
pytest
pytest-asyncio
flake8
aiosqlite
msgpack
